requirements.txt
transformers
torch
streamlit
llama-cpp-python
huggingface_hub